fix: change folder to patient repository
