+++ src/app/modules/Patient/module/index.ts
@@ -1,7 +1,7 @@
 import prisma from '../../../shared/database/prisma';
+import PatientRepositories from '../../../shared/database/repositories/patient';
 import ValidatePatientOwnershipService from '../../../shared/services/ValidatePatientOwnership';
 import PatientController, { optionsController } from '../controller';
-import PatientRepositories from '../repositories/patient/PatientRepositories';
 import CreatePatientService from '../services/Create';
 import DeleteService from '../services/Delete';
 import FindAllPatientService from '../services/FindAll';
