+++ src/app/shared/database/repositories/patient/index.ts
@@ -1,0 +1,3 @@
+import PatientRepositories from './PatientRepositories';
+
+export default PatientRepositories;
